Squish Java porting bug!
Apparently Java will allow you to override a function using a function of the same
name but a different return type. This was causing getOwner() to return nullptr.
Now the base definition is a pure virtual function so getOwner() is ensured to be
properly implemented.

Also did some cleanup work because you know... you gotta.